weston-init: Fix inoperable Weston desktop
For an image configured with pam and systemd, Weston apps
would fail to run. For example:

$ weston-terminal
failed to connect to Wayland display: No such file or directory
failed to create display: No such file or directory

This fix is based on a proposal by Daniel Diaz to OE-core. The
main part of the fix is the use of a new, manually created tty.

The original proposal called weston directly, but that didn't
seem to work where weston-launch did.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>